requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
sqlalchemy
sqlalchemy-repr